Move Insertion into its own module
